# Wrist Strap Monitor System - Data Analysis.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Wrist Strap Monitor System",
    "description": "",
    "version": "1.0.0"
  },
  "tags": [],
  "paths": {
    "/api/v1/logs/": {
      "get": {
        "summary": "Full-Text Search",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "",
            "required": false,
            "example": "Device 10",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/logs/count": {
      "get": {
        "summary": "Test the Count Endpoint",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "",
            "required": false,
            "example": "Connected",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {},
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}